fix: fix component accoriding to comments
remove unnecessary divs,
move styles from separate .css file to styled-component,
add validation to fist name
